release: 2.3.0 출시

+++ data/schedule/src/androidTest/java/com/practice/schedule/util/TestUtil.kt
@@ -8,7 +8,7 @@ object TestUtil {
     fun createScheduleEntityRoom(count: Int): List<ScheduleEntityRoom> = (0 until count).map {
         ScheduleEntityRoom(
             schoolCode = schoolCode,
-            id = it,
+            id = it.toLong(),
             date = (20220601 + it).toString(),
             eventName = "event $it",
             eventContent = "content $it",
@@ -18,7 +18,7 @@ object TestUtil {
     fun createScheduleEntity(count: Int): List<Schedule> = (0 until count).map {
         Schedule(
             schoolCode = schoolCode,
-            id = it,
+            id = it.toLong(),
             year = 2022,
             month = 6,
             day = 1 + it,
